backend/requirements.txt
fastapi
uvicorn
boto3
python-dotenv
python-multipart
httpx
supabase
rapidfuzz
anthropic
pyyaml
pydantic>=2.0
jsonschema>=4.0